+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Atualizações
 
+## [Versão 3.43.82]
+- Relatório de Quantidade de Turmas por Escola (somente administrador)
+
 ## [Versão 3.42.82]
 - Ficha Individual 1º e 2º Segmento